make generic user funciton and put back mapsto
