+++ src/routes/nationalCenter/handlers.ts
@@ -1,5 +1,7 @@
 import { Request, Response } from 'express';
-import { findAll, findAllNationalCenterUsers } from '../../services/nationalCenters';
+import { findAll } from '../../services/nationalCenters';
+import SCOPES from '../../middleware/scopeConstants';
+import { findAllUsersWithScope } from '../../services/users';
 import handleErrors from '../../lib/apiErrorHandler';
 
 const logContext = 'HANDLERS:NationalCenter';
@@ -11,7 +13,7 @@ export async function getHandler(req: Request, res: Response) {
     // all we look for is site access
     const centers = await findAll();
     // Get all users with write training reports permission.
-    const users = await findAllNationalCenterUsers();
+    const users = await findAllUsersWithScope(SCOPES.READ_WRITE_TRAINING_REPORTS);
     return res.status(200).json({ centers, users });
   } catch (error) {
     return handleErrors(req, res, error, logContext);
